+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/Binding.kt
@@ -446,7 +446,7 @@ internal sealed interface Binding : BaseBinding<IrType, IrTypeKey, IrContextualT
       }
 
     override val nameHint: String
-      get() = error("Should never be called")
+      get() = "${typeKey.type.rawType().name}Multibinding"
 
     override val contextualTypeKey: IrContextualTypeKey = IrContextualTypeKey(typeKey)