+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/DependencyGraphTransformer.kt
@@ -800,7 +800,7 @@ internal class DependencyGraphTransformer(
         throw e
       }
       throw AssertionError(
-        "Code gen exception while processing ${dependencyGraphDeclaration.classIdOrFail}",
+        "Code gen exception while processing ${dependencyGraphDeclaration.classIdOrFail}. ${e.message}",
         e,
       )
     }
@@ -991,7 +991,17 @@ internal class DependencyGraphTransformer(
       }
     }
 
-    node.accessors.forEach { (getter, contextualTypeKey) ->
+    val accessorsToAdd = buildList {
+      addAll(node.accessors)
+      addAll(
+        node.allExtendedNodes.flatMap { (_, extendedNode) ->
+          // Pass down @Multibinds declarations in the same way we do for multibinding providers
+          extendedNode.accessors.filter { it.first.annotations.isMultibinds }
+        }
+      )
+    }
+
+    accessorsToAdd.forEach { (getter, contextualTypeKey) ->
       val multibinds = getter.annotations.multibinds
       val isMultibindingDeclaration = multibinds != null