+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/transformers/ContributesGraphExtensionTest.kt
@@ -402,6 +402,46 @@ class ContributesGraphExtensionTest : MetroCompilerTest() {
   }
 
   @Test
+  fun `contributed graph can inject an empty declared multibinding from parent`() {
+    compile(
+      source(
+        """
+            interface MultiboundType
+            abstract class LoggedInScope
+
+            @Inject
+            class MultiImpl : MultiboundType
+
+            @ContributesTo(AppScope::class)
+            interface MultibindingsModule2 {
+              // Important for @Multibinding to be used for this test's coverage, as opposed to @ElementsIntoSet
+              @Multibinds(allowEmpty = true)
+              fun provideMulti(): Set<@JvmSuppressWildcards MultiboundType>
+            }
+
+            @ContributesGraphExtension(LoggedInScope::class)
+            interface LoggedInGraph {
+              val multi: Set<MultiboundType>
+
+              @ContributesGraphExtension.Factory(AppScope::class)
+              interface Factory {
+                fun createLoggedInGraph(): LoggedInGraph
+              }
+            }
+
+            @DependencyGraph(AppScope::class, isExtendable = true)
+            interface ExampleGraph
+          """
+          .trimIndent()
+      )
+    ) {
+      val graph = ExampleGraph.generatedMetroGraphClass().createGraphWithNoArgs()
+      val loggedInGraph = graph.callFunction<Any>("createLoggedInGraph")
+      assertThat(loggedInGraph.callProperty<Any>("multi")).isNotNull()
+    }
+  }
+
+  @Test
   fun `contributed graph copies scope annotations`() {
     compile(
       source(